+++ frontend/src/views/index/AddressBar.vue
@@ -32,6 +32,7 @@ const { locale, t } = useI18n({
             copied: 'Copied',
             fetchAddressError: 'Mail address credential is invalid or account not exist, it may be network connection issue, please try again later.',
             addressCredential: 'Mail Address Credential',
+            linkWithAddressCredential: 'Open to auto login email link',
             addressCredentialTip: 'Please copy the Mail Address Credential and you can use it to login to your email account.',
             userLogin: 'User Login',
         },
@@ -43,6 +44,7 @@ const { locale, t } = useI18n({
             copied: '已复制',
             fetchAddressError: '邮箱地址凭证无效或邮箱地址不存在，也可能是网络连接异常，请稍后再尝试。',
             addressCredential: '邮箱地址凭证',
+            linkWithAddressCredential: '打开即可自动登录邮箱的链接',
             addressCredentialTip: '请复制邮箱地址凭证，你可以使用它登录你的邮箱。',
             userLogin: '用户登录',
         }
@@ -71,6 +73,10 @@ const copy = async () => {
     } catch (e) {
         message.error(e.message || "error");
     }
+}
+
+const getUrlWithJwt = () => {
+    return `${window.location.origin}/?jwt=${jwt.value}`
 }
 
 const onUserLogin = async () => {
@@ -140,8 +146,17 @@ onMounted(async () => {
             <span>
                 <p>{{ t("addressCredentialTip") }}</p>
             </span>
-            <n-card :bordered="false" embedded>
+            <n-card embedded>
                 <b>{{ jwt }}</b>
+            </n-card>
+            <n-card embedded>
+                <n-collapse>
+                    <n-collapse-item :title='t("linkWithAddressCredential")'>
+                        <n-card embedded>
+                            <b>{{ getUrlWithJwt() }}</b>
+                        </n-card>
+                    </n-collapse-item>
+                </n-collapse>
             </n-card>
         </n-modal>
     </div>