+++ worker/src/admin_api/admin_user_api.ts
@@ -5,6 +5,7 @@ import { getJsonSetting, saveSetting, checkUserPassword, getDomains, getUserRole
 import { UserSettings, GeoData, UserInfo } from "../models";
 import { handleListQuery } from '../common'
 import { HonoCustomType } from '../types';
+import UserBindAddressModule from '../user_api/bind_address';
 
 export default {
     getSetting: async (c: Context<HonoCustomType>) => {
@@ -144,22 +145,12 @@ export default {
         }
         return c.json({ success: true })
     },
+    bindAddress: async (c: Context<HonoCustomType>) => {
+        const { user_id, address_id } = await c.req.json();
+        return await UserBindAddressModule.bindByID(c, user_id, address_id);
+    },
     getBindedAddresses: async (c: Context<HonoCustomType>) => {
         const { user_id } = c.req.param();
-        if (!user_id) return c.text("Invalid user_id", 400);
-        // select binded address
-        const { results } = await c.env.DB.prepare(
-            `SELECT a.*,`
-            + ` (SELECT COUNT(*) FROM raw_mails WHERE address = a.name) AS mail_count,`
-            + ` (SELECT COUNT(*) FROM sendbox WHERE address = a.name) AS send_count`
-            + ` FROM address a `
-            + ` JOIN users_address ua `
-            + ` ON ua.address_id = a.id `
-            + ` WHERE ua.user_id = ?`
-            + ` ORDER BY a.id DESC`
-        ).bind(user_id).all();
-        return c.json({
-            results: results,
-        })
+        return await UserBindAddressModule.getBindedAddressesById(c, user_id);
     },
 }