+++ worker/src/user_api/index.ts
@@ -6,12 +6,17 @@ import user from './user';
 import bind_address from './bind_address';
 import passkey from './passkey';
 import oauth2 from './oauth2';
+import user_mail_api from './user_mail_api';
 
 export const api = new Hono<HonoCustomType>();
 
 // settings api
 api.get('/user_api/open_settings', settings.openSettings);
 api.get('/user_api/settings', settings.settings);
+
+// mail api
+api.get('/user_api/mails', user_mail_api.getMails);
+api.delete('/user_api/mails/:id', user_mail_api.deleteMail);
 
 // user api
 api.post('/user_api/login', user.login);