+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - feat: |UI| 支持 URL jwt 参数自动登录邮箱，jwt 参数会覆盖浏览器中的 jwt
 - fix: |CleanUP| 修复清理邮件时，清理时间超过 30 天报错的 bug
 - feat: admin 用户管理页面: 增加 用户地址查看功能
+- feat: | S3 附件| 增加 S3 附件删除功能
 
 ## v0.9.0