+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - feat: admin 用户管理页面: 增加 用户地址查看功能
 - feat: | S3 附件| 增加 S3 附件删除功能
 - feat: | Admin API| 增加 admin 绑定用户和地址的 api
+- feat: | Oauth2 | Oatuh2 获取用户信息时，支持 `JSONPATH` 表达式
 
 ## v0.9.0